+++ tests/configs/config_ds000248.py
@@ -37,6 +37,7 @@ def noise_cov(bp):
 spatial_filter = 'ssp'
 n_proj_eog = dict(n_mag=1, n_grad=1, n_eeg=1)
 n_proj_ecg = dict(n_mag=1, n_grad=1, n_eeg=0)
+ssp_meg = 'combined'
 ecg_proj_from_average = True
 eog_proj_from_average = False