+++ docs/source/changes.md
@@ -250,6 +250,10 @@ authors:
 - Patch information is now incorporated when computing surface source spaces,
   which should slightly improve the surface normals
   ({{ gh(588) }} by {{ authors.larsoner }})
+- Add [`ssp_meg`][config.ssp_meg] option for MEG SSP computation. This
+  defaults to `'auto'`, which will use `ssp_meg='combined'` for SSP computation
+  when Maxwell filtering is used.
+  ({{ gh(595) }} by {{ authors.larsoner }})
 
 ### Code health